Always provide CellIdentity in NetworkRegistrationInfo
Always return the CellIdentity, if available in NRI. Even
when the caller does not have / renounces location usage,
there is useful info in the CellIdentity, in particular the
PLMN-ID and the band(s) that the cell is using. Because
band info is supplied by the cell's broadcast info, this is
the only place where it can be retrieved.

Bug: 357159744
Test: atest NetworkRegistrationInfoTest#testSanitizeLocationInfo
Test: basic sanity test (voice, messaging, data)
Flag: EXEMPT bugfix
Change-Id: I0a424fdc67d1452a4a11a20980672841543c16ac