[Fold/unfold] Wait for the present fence of Shell transition's start transaction
We were not waiting for the present fence of
a surface control transaction, we were just listening
for the transaction completion callback before unblocking
the screen after folding/unfolding.

The screen unblocking is done by updating the
brightness which doesn't go through surface control
transactions, meaning that it could be updated earlier
than a transaction that is applied after brightness
update request => it wasn't guaranteed that the buffer
will be presented when the completed callback is fired,
which could cause a flicker with the old content.

Test: manual fold/unfolds
Bug: 370719724
Flag: com.android.window.flags.wait_for_present_fence_on_display_switch
Change-Id: I0f1b5ccca926ab889dba4d9f9481c79c6ca7b8b6