Update surface for the HardwareRenderer.
When using lockHardwareCanvas() on Surface, if the device is rotated,
then the changed size information is not set to the HardwareRenderer.
So as the area to be drawn changes, there is a problem where only some areas are drawn.
The problem occurs only on Skiagl, but it is also necessary to make this correction for all renderer like Skiavk.

Test: make
Flag: EXEMPT bugfix
Fixes: 386837581

Change-Id: I31ebff7e982d3ae8ca8ccea77764bbcf27c26b05