[1/N] Reparent desks on display disconnect.
When a display containing a desk is disconnected, start a transition to reparent the desk to
the display area of its destination display. TaskDisplayArea will set the
desk root task to top if it was on top of the focused display or bottom if not.

On a destination display that supports desks, ensure the reparented desk is activated if it was on top of the focused display. Otherwise, move it to the display below
whatever is currently active. Whichever desk moves to the bottom will
then be deactivated.

Bug: 391652399
Test: Manual
Flag: com.android.window.flags.enable_display_disconnect_interaction
Change-Id: I6ec4b9d305a6bd9497450328dad36bc53179ee7c