AudioDeviceInventory: fix Spatial Audio for LE broadcast.
LE broadcast devices are virtual devices that should not have their own
SA state. Also head tracking does not make sense in a broadcast situation.

The SA state for a broadcast device is now derived from the SA state of
the corresponding unicast device except for HT which is always disabled.
The state is not persisted and cannot be changed by Spatializer APIs.

Bug: 380151062
Test: make
Flag: EXEMPT bug fix.
Change-Id: I6aeee91993d1755b100d01919383e33aa009cd9d