Making the Bubble task to continue launch next as bubble
This CL also fixes the issue of unable launching the apps that
use Task trampoline into Bubble. That's because the new Bubble Task
inherits the windowingMode and bounds from the trampoline Bubble
Task, which allows the launch cookie to be propagated.

However, we still need to ensure the bubble task is always-on-top,
because it may not be the initial task if the app trampolines
to a new task.

Bug: 388651207
Test: wm presubmit
Flag: com.android.wm.shell.enable_create_any_bubble
Change-Id: Ia69d449981d2e718f512e052691622b713baa632