Desks: Incompatible Tasks (1/2) - Merge isDesktopModeShowing()/isAnyDeskActive()
isDesktopModeShowing() is the same as isAnyDeskActive() but with extra
considerations for top-transparent-fullscreen tasks. This extra
information is available in the repository, so it can be made part of
isAnyDeskActive() there.

For multiple desks, top-transparent-fullscreen checks aren't needed
anymore because a desk's active state is not tied to freeform task
visibility or top-transparent-fullscreen task existence, and is instead
set/unset explicitly.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 398295550
Test: launch incompatible tasks in desktop, verify they remain
fullscreen and activate/deactivate the desktop based on transparency

Change-Id: I49fd8edc2c03242ce80b3cfcc140b93d3438708b