Cache ViewConfiguration instances per device
Currently, ViewConfiguration instances are cached based on
display density. This means that displays belonging to different
devices can end up using the same ViewConfiguration instance, if
they have the same density (this can happen when an app is having
multiple activities on multiple devices at the same time, such as
one activity on the phone and another on Chromebook). This change
ensures that we make the caching per device by constructing a key
made of deviceId (associated with the context) and the display
density.

Test: atest ViewConfigurationTest
Bug: 405726390
Flag: android.companion.virtualdevice.flags.viewconfiguration_apis
Change-Id: Ie5953150af530c6f48ea57f15b011da097f353bf