Customize restart dialog icon for restart handle menu
The current icon for the restart dialog is intended only for size
compat mode, so we need to replace it with a normal "restart" icon
when the dialog is requested by the restart handle menu (as it's
available to non-SCM apps too).

This change simply replaces the src drawable of the ImageView when
isRestartMenuEnabledForDisplayMove() is true.

Flag: com.android.window.flags.enable_restart_menu_for_connected_displays
Bug: 406072124
Test: RestartDialogWindowManagerTest
Change-Id: I513ac27e6ee38af7c61dedf2a2706930b05ec46d