[um] Suppress StrictMode warnings for user data destruction when the user is locked
The SystemServer should be able to destroy a user data directory while
the user is locked, without triggering the StrictMode warnings. This CL
temporarily disable the StrictMode detection for
DETECT_VM_CREDENTIAL_PROTECTED_WHILE_LOCKED before destroying the user
data, and restores to the original policy value afterwards.

Notice that this CL is changing a VmPolicy not a ThreadPolicy.

BUG: 347054955
FIXES: 347054955
FLAG: EXEMPT bug fix
Test: atest FrameworksCoreTests:android.os.StrictModeTest

Change-Id: Idfff0e4f28b5aff76067d510a0be0ffd1116e88f