Pass reject count in batched noteOp call
In batched noteOp calls we pass the aggregated count to the system
server. If the op access is allowed, it will be access count; if the
access is denied, it will be reject count. Currently we don't pass
reject count to HistoricalRegistry for tracking, the default count is 1.

When aggregating noted ops on the client side, we need to account for op
mode, since in rare cases the mode could change during 1 second of
batching time.

Bug: 406281478
Test: atest android.app.appops.cts.HistoricalAppopsTest
Flag: EXEMPT BUG_FIX
Change-Id: I9ea57074cd0024ae25dad65ff287333e5008562e