Remove flag android.adaptiveauth.report_biometric_auth_attempts
Since this flag was launched, remove it and make the corresponding code
unconditional.

Note that it's necessary to remove the adaptiveauth flags library rather
than leave it empty, since the build system errors out when a flags
library is empty.  It is planned that future flags in this area will go
in a different location such as
./core/java/android/security/flags.aconfig.

Bug: 285053096
Test: presubmit
Flag: EXEMPT removing android.adaptiveauth.report_biometric_auth_attempts
Change-Id: Ia573f37e855d61ab5ded885b2a665c1af4fd306c