Reset bubble bar expanded view clip after collapse
When the bubble bar is expanded and the IME is hidden it's possible
to cause it to collapse without hiding the IME first, e.g. by
pulling down the notification shade and going to settings. This
change resets the clipping of the expanded view so that it shows
up correctly the next time it is expanded.

Flag: EXEMPT bug fix
Fixes: 391353155
Test: manual
       - add bubble to bubble bar
       - expand and show IME
       - pull down the notification shade
       - tap on Settings
       - swipe home
       - expand bubble bar
       - observe expanded view is not clipped
Change-Id: Ibde608cef0455f0ef1cca012cba975f1dda7983c