Move getResources() off of boot critical path
When starting VoiceInteractionManagerService, there is a getResources()
call that blocks boot for 150ms. This happens when the service's
RoleObserver is instantiated. This change updates this work to happen in
a separate thread, with results being returned in a future.

This change also updates SystemServerInitThreadPool to allow submitted
tasks to return results.

Test: verify in traces that resources are loaded in a separate thread
Bug: 406841419
Flag: android.server.voiceinteractionmanagerservice_get_resources_in_init_thread
Change-Id: I029a09e7d1bb6af1aeebcd65bfefa3a9a46540fe