Correctly update preview clock size when requested
Previously we were only updating the smartspace position, and the clock
preview binder was not listening to the appropriate signal. This fixes
that issue by moving the source of the override flow. This bug was
latent until we enabled the new wallpaper flow in picker.

As part of this, I've started a refactor of how our preview data is
organized to better match modern architecture guidelines. More work here
is needed as much of the data is still accessed directly by, or set
directly on KeyguardPreviewRenderer itself. Ideally this could be
treated much more like a View or Binder instead of doing so much
directly.

Bug: 375207082
Bug: 407131820
Test: Manually checked preview behavior
Flag: build.RELEASE_PACKAGE_MAGIC_PORTRAIT_WALLPAPERS
Change-Id: Idc8b3c736b62ff9bbae10d693e3062566744402d