Refactoring KeyguardQuickAffordanceHapticViewModel.
The refactor simplifies the logic to play haptics by removing the need
to rely on Flows. Now, the view model is only responsible of playing
haptics when a quick affordance toggles after being long-pressed. The
logic to play haptics when a quick affordance launches is now delegated
to the KeyguardQuickAffordanceInteractor.

Test: modified and added Unit tests
Test: manual. Verified that correct haptics play when toggling shortcuts
  in the lockscreen, as well as long-pressing to launch activities.
Test: manual. The same manual tests were verified with flexiglass on.
Flag: com.android.systemui.msdl_feedback
Bug: 394330308
Change-Id: I19a189c0b94ed0813c46da9f08297cb9777764b0