Exclude ViewPager2 from default transition in KeyguardClockViewBinder.
When `TransitionManager.beginDelayedTransition` is called without
passing a transition, the default `AutoTransition()` is used. The
`AutoTransition()` contains transitions that subclass the `Visibility`
transition.

`Visibility.onDisappear` invokes the `onViewAdded` method, which isn't
supported by `ViewPager2`. As a result, this leads to the "ViewPager2
does not support direct child views" crash.

This is the only `Visibility` transition I found in the keyguard code
that doesn't exclude the `ViewPager2`. So hopefully it fixes the crash.

Flag: EXEMPT bugfix
Fix: 387398068
Test: See #comment21 and #comment24 in the bug for steps
Test: Crash trends
Change-Id: I896023db2b8181cd73e2cc76a470b3a218c12f0e