Apply parallax on offscreen tap animation
Previously, when the user tapped an offscreen app, the resulting "come back onscreen" animation did not have parallax effects applied. This created a slight (undesirable) shifting effect on the left/top side, due to the uneven contraction/expansion of surfaces as they moved on- and offscreen.

Fixed by adding a mid-animation state to SplitState, and applying different parallax effects during this state. This also contributes to an overall effort to have SplitState be the new source of truth regarding what is going on split -- including all mid-animation states.

Note that, interestingly, we need different parallax effects depending on if the user is dragging the divider from 10:90 to 90:10 or if the system is animating the divider along the same path. This is because in the user-controlled case, we don't know where the user plans to drag to, so we need to add a center-squeeze effect in case the user is going to 50:50.

Fixes: 402609970
Flag: com.android.wm.shell.enable_flexible_two_app_split
Test: Manually verified that the new animation looks better, and does not affect other transitions like user-controlled, launch_adjacent, etc. The landscape phone case also looks a lot better.
Change-Id: I89c7027071d79a366cabaddf2117eed253c87ea9