Don't collapse Bubbles when task opens on a different display
The current implementation collapses bubbles when any task opening
transition is detected outside of a bubble. This leads to also
collapsing the bubbles even when the user starts an app on a
different display, and there is no overlap or visual relation
between the bubbles stack and the new task.

This change adds a check on which display the task is opening,
and skips collapsing if it is not where the bubbles are currently
showing.

Bug: 408249958
Test: atest com.android.wm.shell.bubbles.BubblesTransitionObserverTest
Flag: EXEMPT bugfix
Change-Id: I05c7acea928143e45bbf9d00698c82485726ba85