Complete removal of device policy from UserManagerInternal
DEPRECATE_USERMANAGERINTERNAL_DEVICEPOLICY_FLAG defaults to true, and no
one appears to be overriding it via DeviceConfig.  Therefore, all the
code that is conditional on this flag being false is unused.  Remove it.
Then remove the UserManagerInternal methods that were only "used" by
this unused code, and the mIsDeviceManaged and mIsUserManaged fields in
UserManagerService which existed only to support these methods.

Bug: 258213147
Flag: EXEMPT removing already-launched flag
Test: atest Frameworks{,Mocking}ServicesTests:com.android.server.{devicepolicy,locksettings}
Change-Id: I06ce80ad2eba5ad2c04ed73055c4d2289cbc3be6