Fix installing of DEVICE_POLICY_MANAGEMENT Role holder in work profile
When a work profile is created, the 'DEVICE_POLICY_MANAGEMENT' role
holder app must be installed in that new profile.
To do that, the code must figure out who holds that role *on the current
user*, so that same app can be installed in the work profile.

The old code erroneously checked who had the role in the *work profile*.
This only worked because CloudDPC was always already installed.

Bug: 408993396
Flag: EXEMPT bugfix
Test: Manually created work account and verified it didn't crash.
Change-Id: I8e3159734200bfd592591566a3bd57e0d9dc3851