Clear size compat config when resizing to large screen
The CompatDisplayInsets could be created when the activity is
on a small screen. If the activity is moved or the display is
resized to large screen, then the compat config should be cleared
to fit container size if universal resizable is enabled.

For example:
 Fold to unfold(large screen): clear size compat info
 Unfold to fold(compact screen): create size compat info

Fix: 407952252
Flag: com.android.window.flags.universal_resizable_by_default
Test: SizeCompatTests#testUniversalResizeableByDefault
Change-Id: I4da10de140c8bdf565e97239e9ea3901c293ea3d