Desks: Use a single transition for desktop->split flows
Using two separate transitions (the one started by requestSplitSelect
with the WCT desktop provides + the recents ones) causes them to be
collected into a single transition that keeps the token of the first and
discards the token of the second. This caused Recents to be unable to
identify a controller to handle the transition since it expected the
token of its own startTransition.

In general, starting the first desktop->split-select transitions
shouldn't be needed in the first place, so this CL combines the two WCTs
and lets it be applied on the recents start.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 409639676
Test: enter split from desktop handle menu, verify it enters split
select

Change-Id: I3195768c01ae3c869e7203d0a51072cf59105d24