Desks: Do not assume a desk exists when checking for tiled tasks
DesktopRepository#getLeftTiledTask and #getRightTiledTask are invoked
whenever a window decoration is created, even if the decor is for a
fullscreen task and the device has never entered desktop windowing.
This removes the assumption that a desk will exist in the repository, as
that is not always the case, such as in touch-first displays.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Fix: 410047482
Test: atest ActivityLifecycleTopResumedStateTests
Change-Id: I18587240d41f7c0c916ca3cafa3d2afbc6fa973f