power: Move setting `sys.shutdown.requested` earlier
The `init` process will start monitoring `sys.shutdown.requested` and
kick off a watchdog to make sure that the "Shutting down..." dialog
can't hang forever. Move the setting of this property earlier, but
still after we've committed to shutting down. Specifically, move it to
be before we create the dialog and block the user from doing anything
else.

Bug: 409835922
Test: With ag/32996886, various shutdown types trigger proper watchdog
Flag: EXEMPT minor change
Change-Id: Idadb2acbaac00c489a3c27e2d1c8746728dd6a4b