Replaces BP UDFPS guidance disruptive accessibility announcements
Replaces disruptive accessibility announcements with live accessibility
region for BiometricPrompt. Because these Talkback announcements are
propagated via content description changes, we have to clear the content
description each time between UDFPS guidance message to allow sending
the same message multiple times in a row. We also have to mark the
accessibility overlay not important for accessibility when the user
lifts their finger, in order to prevent linear navigation to an overlay
with a stale or unlabelled view.

Flag: NONE bug fix
Fixes: 407619360
Fixes: 404938119
Test: atest PromptViewModelTest
Test: atest BiometricPromptUdfpsAccessibilityOverlayViewModelTest
Change-Id: If4d428009070eaff08b1aeded0b95632f591637a