Desks: Do not assume moveToFront means moving to desk
DesktopTasksController#moveToFront may be used on fullscreen or split
tasks that aren't focused (on display or globally) to bring them to the
front when the App Handle is touched. Do not assume that means they need
to be moved into a default desk and that a desk needs to be deactivated.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 410180972
Test: open task, make it unfocused, touch the app handle - no crash
Change-Id: I18ce831984d6e869f89af5ba5d2db2ae97a16f13