TIF: update throw exception to log exception in servicestate
When an tis that is hardware is not allowed for kids, it could lead to
IllegalStateException when querying servicestate. This is because the
hardware input is added on owner session as it is allowed for the owner.
As we intent to share hardware inputs across profiles, this should not
happen when allowlist is finalized. For now, we should indicate the
exception with log instead of throw the exception, leading to the system
to crash unncessarily.

Test: manual
Bug: b/408900599
Flag: EXEMPT minor fix
Change-Id: Iadf1fecb962969f6e5a01fb3fa53fce29ace8348