Fade the bouncer when the dream starts
This is not a common CUJ, but the dream can start while the primary
bouncer is showing if the user presses power or the screen times out
while the user's "When to show" condition for the dream is active.
Since the dream is an activity, it starts underneath the bouncer and
the animation looks bad.

This change adds an alpha animation to the bouncer when the dream
starts.

Bug: 409720192
Test: atest FromPrimaryBouncerTransitionInteractorTest
Flag: NONE, fixing long-standing bug
Change-Id: I9b4bf6ba9560d6beabd77c591718520e55967e41