Ensure dispose is always called on looper thread -- try 2
The InputEventReceiver::dispose should only be called on the looper
thread. This is already documented in the class, but not enforce.

Enforce this here, since this will become critical in subsequent patches
where we are relying on the ability to consume upon dispose.

This is the second attempt, after finding a bunch of issues in the postsubmit.

Bug: 376713684
Test: atest InputEventSenderAndReceiverTest
Flag: EXEMPT refactor
Change-Id: I2484f17491cf5f9afba5ec68f759595ccf3329ce