Update height of QQS in shade when media appears
The previous fix only updated when media disappeared. This was done to
minimize impact of change, as the issue was only seen when media
disappeared. However, we have observed cases (in tests) of media
appearing when the shade is open and the height not updating.

Test: atest android.platform.test.scenario.sysui.media
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Bug: 406248870
Change-Id: I1075d8f12acf3d3dfa2fd039e54ce7a5f2d54973