Playing dismiss haptics when flinging to dismiss but without detaching.
The dismiss haptics that play when a magnetic notification detaches
should also play in the event that the user flings to dismiss a
notification, but the notitication hasn't magnetically detached. This
change adds this feedback and makes the dismiss cases fully consistent
in terms of haptic feedback.

Test: added Unit tests
Test: manual. Verified that the dismiss haptics are felt when the user
  flings a notification to dismiss it without previously dragging it
  beyond the magnetic detach threshold.
Flag: com.android.systemui.magnetic_notification_swipes
Bug: 411389706
Change-Id: I28f7c754f6489c837df4be876a71154c1d53704c