Update the hardcoded value for the const CORE_DEFER_UNTIL_ACTIVE.
This is already set to true when broadcast constants are
initialied. So, this only impacts tests that does not initialize
broadcast constants.

This makes the tests behavior consistent which results in one of
the tests failing, as this test relies on running after
'testCalculateDeferUntilActive_System' which may not always be
the case.

Bug: 411309776
Test: atest services/tests/mockingservicestests/src/com/android/server/am/BroadcastRecordTest.java
Flag: EXEMPT bufix
Change-Id: Ia4d2a34343bb28c782f27f6e41088fa8cdf41dca