Desks: Do not trigger taskbar anim on recents-finish out of desk
DesktopModeEnterExitTransitionListener#onExitDesktopModeTransitionStarted
is meant to notify the taskbar of a transition to exit desktop mode that
has just started but is not yet animated so that the taskbar animation
to hide itself can trigger (roughly) in sync with the transition.

I34b9a581db3af5e160e32b005d35742089f55cb7 added a change to trigger a
desk deactivation, and thus a |onExitDesktopModeTransitionStarted| call
in |onRecentsInDesktopAnimationFinishing| because with multiple desks,
a desk may need to be deactivated if finishing recents to home or
another non-desk task. However, the |onExitDesktopModeTransitionStarted|
call is not needed here as the taskbar does not need to be animated, and
it would also be too late to animate it, since this is called on
transit-finish, not on transit start. This change removes that call from
this path.

Also guards all calls to this interface behind a check for both the
frontend and backend flags, because this interface is deprecated when
both flags are enabled.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Flag: com.android.window.flags.enable_multiple_desktops_frontend
Bug: 411597362
Test: ran systemui-desktopmode-enter-1-jank-suite locally
(EnterDesktopViaStaticDesktopOverviewTaskMicrobenchmark) and verify
latency increase is fixed.
Test: atest DesktopTasksControllerTest

Change-Id: I10188542e4a6575b1eebb202a1c6ea9cc83998d2