Desks: Force freeform-incompatible tasks to fullscreen
When not in desktop, but the TDA is freeform and an incompatible task
launches in freeform mode by default, it should be forced out into
fullscreen mode otherwise it will remain in freeform but won't belong to
any desk. This can later lead to a NPE trying to add the freeform task
to the DesktopRepository that has no desks in it.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 410731810
Test: test launching incompatible freeform task in touch-first display
that has 0 desks - verify it launches in fullscreen and there is no
crash

Change-Id: I562adba92c69af9a0f894601a723b0514ef2340d