Do not transfer the starting window if it is attached to client.
Once transferSplashScreen occurs, the starting window cannot be
forwarded to any other activity.

Also, fix a flicker that occurs after onSplashScreenAttachComplete.
When this callback is received, the splash screen window's animation
leash is already hidden, and the window should be removed immediately.
Therefore, clearing the leash at that moment is unnecessary and may
cause flicker because the window is still visible, keeps the leash
hidden and it will be removed automatically when the window is removed.

Flag: EXEMPT bugfix
Bug: 402644135
Test: launch Activity A which register setOnExitAnimationListener, then
launch Activity B from A, verify the splash screen window won't transfer
the B.

Change-Id: I5604603b80afd9f6e1425645414e26e0e1b93311