Reduce unnecessary task initialization from empty info
When adding an activity to a new task, setIntent and setMinDimensions
will be called from setNewTask > reuseOrCreateTask> reuseAsLeafTask
with a valid ActivityInfo. So it is unnecessary to initialize the
task with an empty ActivityInfo.

Set Task.Builder#mResizeMode to RESIZE_MODE_RESIZEABLE because
that is the default value of an empty ActivityInfo.

Bug: 297502610
Flag: EXEMPT Remove duplicated invocation
Test: TaskTests
Change-Id: Ie2e09f553932c20ef7e6144f33df3954745df47e