Fix hub to edit mode transition
Previously while transitioning from hub to edit mode, the communal scene
stays on as keyguard transitions to GONE. This was to coordiate a
smooth transition between the two states. But this breaks the state
syncing between communal scene layout and KTF.

This change makes it so that KTF does not immediately transition to
GONE, but let the CommunalSceneTransitionInteractor handle it when the
communal scene is ready to transition away.

Test: atest FromPrimaryBouncerTransitionInteractorTest
Test: atest FromAlternateBouncerTransitionInteractorTest
Test: manually tested the transition between hub and edit mode
Test: b/411422599 does not reproduce
Fix: 411422599
Flag: com.android.systemui.hub_edit_mode_transition
Change-Id: Ib4b3be1504d352a26f2ac220722fe1c3be157394