1.3 Extract window selection to the wrapper.
This is the last step before wrapping the data in an API-agnostic
class that will allow us to use the impl with both the legacy and
Shell API.

Trying to make each of the CLs as small as possible to keep them
digestible and low risk. For the refactor plan see
go/animlib-shell-refactor-plan.

Bug: 397180418
Flag: EXEMPT simple refactor
Test: atest ActivityTransitionAnimatorTest + manual
Change-Id: Ib1a4cf3e2f097db4ec950237dec0564ea3453271