Protect against null mDreamToken.
A wakelock wrapped handler was added previously to help protect access
to mDreamToken. This led to an increase in binder calls due to the
wakelock acquisition. Even though this was adjusted to account for
operations already on the main thread, this is a deviation from the
original behavior execution proceeded on any thread. This change
restores the original behavior and instead now checks the nullability
of mDreamToken before using it.

Test: atest DreamServiceTest
Fixes: 408928049
Fixes: 408393208
Flag: EXEMPT bugfix
Change-Id: If7ea441b808a467c3f777f7ba4b36e319e96a411