[PiP on CD] Reduce overlap of drag region with PiP window.
Recall: http://recall/clips/7719aba5-cc34-4072-9f4f-0b5532209201

The overlap between drag-to-resize region with the PiP window is causing
single click input events on the PiP menu buttons (specifically top left
Settings and top right dismiss) to not receive input. Although this is a
general PiP on Desktop Windowing issue, it is much more prominent when a
mouse is used, and therefore more relevant to CD use case.

This CL reduces the overlap for the drag region so that input to those
buttons are not blocked.

Bug: 399196269
Test: atest WMShellUnitTests
Test: Manual - verify drag-to-resize and button clicks are WAI with
Desktop PiP
Flag: com.android.window.flags.enable_connected_displays_pip

Change-Id: Ie59351b3db8457cae3d88d0fac020100bc614de1