Able to remove starting window right after onTransitionReady
If a starting window has participated in a transition but is ready to
be removed before the transition is ready, defer the remove request
right after onTransitionReady sent to Shell. This allows Shell to
handle the removeStartingWindow request earlier and ensures that flags
related to the starting window are correct within TransitionInfo.

Bug: 415696165
Flag: com.android.window.flags.remove_starting_in_transition
Test: atest FlickerTestsActivityEmbedding
Test: atest StartingWindowControllerTests
Change-Id: If0110ab134177c1fd1720aa6238647c6fb7d2786