Set per display status bar window titles
To make it easier to debug in web-hv or hsv, this adds per display
window titles.

For the default display, the title will be as before "StatusBar". For
non default displays, the title will have the display id suffix.
For example, if the display id is "123", the title will be
"StatusBar123".

Test: StatusBarWindowControllerImplTest.kt
Flag: com.android.systemui.shared.status_bar_connected_displays
Bug: 362720336
Change-Id: Ifc9c536c341c165e648f0d7fae70f56ea070d33a