[MQ] Fix IMAX_SDR can't be found
The reason IMAX_SDR can't be found is that there's an extra space in the
name.

Along with the fix for the behivor of notify the HAL about picture
profile. The problem is we only store orignal handle in the
mPictureProfileForHal set. So when non-orignal picture profile get
update, the HAL can't get notify.

Bug: 416211509
Bug: 409274226
Test: mmm
Flag: EXEMPT bugfix
Change-Id: Ib7cdceb41dbf5f3bcaa05ea4bb064d83126d3b8e