Delegate mirror display permission check to VDM.
We'll need to do app compat check before checking for the permission
and we'll need the VD owner info for this, so it has to be done in VDM

Any future use cases for bypassing the permission due to user consent
need to be handled in VDM as well.

This also cleans up the DMS tests.

Bug: 370657575
Test: presubmit
Flag: EXEMPT no-op
Change-Id: Ie452b08933b564c0faf3b7a6329d019dd48b878c