Implement attachChildSurfaceToTask() in RootTaskDesksOrganizer
Now that DW tasks are orgnized by the desk organizer, we need to
implement attachChildSurfaceToTask() in RootTaskDesksOrganizer so
CompatUI can add dialogs etc to DW tasks. (Most of the app-compat
features are disabled in DW, but the new restart handle menu is
an exception.)

attachChildSurfaceToTask() needs a task's leash provided by
onTaskAppeared(), so we need to maintain leaf leashes in the
organizer as all other organizers do.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 416126996
Test: Manual. "Optimize View" no longer crashes.
Change-Id: Ic766c1a51551a315a38e475524e090e455618623