Refactor: Use supportsSwitchTo() in UserManagerService
Switch from `supportsSwitchToByUser()` to `supportsSwitchTo()` for
determining user switchability.

The `supportsSwitchTo()` method is more suitable for framework-level checks as other method supportsSwitchToByUser() description indicates that it should be true if this user can be switched to by the end user through the UI. This aligns better
with the framework's requirements. when I added this logic I didn't pay attention so fixing it now.

Bug: 418022521
Flag: EXEMPT refactoring
Test: manually build and deploy in DUT
Change-Id: I1d4c7db36ecda75a46b28bd687f58740919e2de3