Do not pass indices between drag state and content list state.
Currently, the position of a items in the lazy grid are passed to the
content state in order to perform operations on the object. This can
lead to inconsistencies as it is not guaranteed these two
representations are kept entirely in sync. For example, a placeholder
might be removed or an object adjusted before an operation may occur.

This changelist addresses this by using the item key as a stable
identifier. This value can be used to look up the position of the item
in each list representation.

Fixes: 417338925
Test: manual - ensured drag and drop operations worked without crashing
Flag: EXEMPT bugfix
Change-Id: Ic6681867066fafa5c12c083d1c700e5ead9c3de5