Fix a crash that occurs when drag-reordering widgets.
The crash occurs because GridDragDropState adds a Responsive
CommunalContentSize to the end of the list of items to support dragging
to the end, but later code attempts to cast it to a FixedSize
CommunalContentSize. The fix conditionally creates the correct class
based on whether the communal_responsive_grid flag is enabled or not.

Fixes: 417999806
Test: manually by dragging with all combinations of
communal_responsive_grid and glanceable_hub_v2 either enabled or
disabled.
Flag: EXEMPT bugfix

Change-Id: I515e61055c43fd262672bddd67a949e53d5101f7