DisplayTopology.rearrange: more corner attachment
Simplify the criteria by which displays are attached to each other since
they resulted in intended corner attachments not being used and distant
displays clamping together instead.

The prior criteria were written with the motivation to maximize the
length of the clamp, which penalized corner attachments. This is not
necessary because the input manager graph is generated elsewhere, and it
detects large clamp lengths reliably and even allows cycles, whereas
DisplayTopology does not.

Test: DisplayTopologyTest
Test: In topology pane in connected displays settings, attach three displays together via corners as in b/394355269 where the root only has one child. Try many variations to attempt to trigger the bug.
Bug: b/394355269
Flag: com.android.server.display.feature.flags.display_topology
Change-Id: I8463d14034faacdb1deee8b665003bd3f6670fcc