Remove text selection toolbar timeout
A timeout exists that assumes that if the toolbar hasn't been dismissed
by the client after a fixed 10 minutes then the toolbar is invalid and
can be evicted from the cache. It is more correct to clear when the
client no longer exists.

It was also found that quick tests were able to sometimes trigger the
toolbar to attempt to show with a null ViewRootImpl; it's unclear how
this happened but the view is not in a state to show the toolbar so exit
early.

Test: CtsWidgetTestCases
Fixes: 413215561
Bug: 363318732
Flag: android.permission.flags.system_selection_toolbar_enabled
Change-Id: Ic74ca4f029d9aaa7a44fac4d9927fdc4f5bd8e2b