Stop greying out icons in group children
In an effort to reduce duplicate information in groups, we would make the color of group icons grey and only show it in the header if it was the same. With the new design and themes though, the grey looks more out of place, and small icons can only be used by system apps anyway, which primarily use the default color. We can therefore remove this logic.

Fix: 414555351
Test: manual, temporarily made all apps show small icons to test
Flag: android.app.notifications_redesign_app_icons
Change-Id: Ibf5c11dd0be1e58221822352305c89da04ea43e3