Screen Capture: Create RadioIconButtonGroup component
Creates a component which hosts a variable number of icon buttons where
any single icon button is selected at a time. The selected and
unselected button colors can be customized via argument. This is
intended to be used for the capture area selection in the toolbar.

Bug: 412722889
Test: Screenshot tests in topic
Flag: com.android.systemui.desktop_screen_capture
Change-Id: Ice184ed60a4602c9c032131d3bd6c602ea0f78b8