Remove leftover wallpaper dim on reboot
Unlike other device effects (AOD supression, grayscale, etc) dimWallpaper survives reboots. Because DefaultDeviceEffectsApplier only applies _changed_ effects, this means that dimWallpaper would be stuck if it was active before the reboot, but inactive afterwards.

Although this condition (mode stopping on reboot) is rather rare, it can happen, for example, if DND was set with a timeout and the timeout expired between power-off and power-on.

Fixes: 419466188
Test: atest DefaultDeviceEffectsApplierTest
Flag: EXEMPT Bugfix
Change-Id: Iaeb4b09b4378310142ec1fefe6f9a7409e48411c