Send intent to settings to open screen lock dialog
Add ScreenLockDialogContract and use the intent it builds to launch the
dialog that promtps the user to set up a screen lock before interacting
with the Private Space.

The dialog is being moved to the Settings app to allow theming.

Bug: 406130727, 407327874
Test: manual and presubmit
Flag: android.multiuser.move_set_screen_lock_dialog_to_settings_app
Change-Id: I466fc310fba22b16f992a6fbcc4000cef83a1d36