Simplify SelectionToolbarManagerService
This service is simply a passthrough to the render service, complex
frameworks to manage events such as user lifecycles are not needed to
be implemented for such a simple functionality.

We also only reference the render service in the system user now since
either process, system ui or android:ui, should be run as the system
user.

Test: CtsWidgetTestCases + Manual
Bug: 363318732
Flag: android.permission.flags.system_selection_toolbar_enabled

Change-Id: I5925387d42a54058ce21f47554378564185e06b6