[43/n] Define LetterboxLifecycleEventFactory
The event triggering some Letterbox Surface lifecycle operation
can come from different sources which are abstracted with
LetterboxLifecycleEventFactory.

MultiLetterboxLifecycleEventFactory implements the same interface
adding a Chain of Responsibility logic which delegates the
creation of the LetterboxLifecycleEvent to the first object
that has all the data to do it.

LetterboxLifecycleEventFactoryUtils provides some utilities to
simplify the test of any LetterboxLifecycleEventFactory
implementation and starts the definition of different
TestInputBuilder to be used when common objects need to be
created (e.g. Change, TaskInfo, etc).

Flag: com.android.window.flags.app_compat_refactoring
Bug: 409043134
Test: atest WMShellUnitTests:MultiLetterboxLifecycleEventFactoryTest

Change-Id: I4c2ad7177a73cd7d507d7a5ec06a7295fcb553dc