Pause transient launch after switching to a translucent task
The activity lifecycle change of transient launch is usually deferred
until the transition is finished to reduce switch latency. It is fine
if transient launch will be invisible because ensureActivitiesVisible
will schedule to stop the activity. But if not, such as moving a
translucent task to top, then the non-top visible home will stay at
resumed state unexpectedly.

Bug: 419630442
Flag: EXEMPT bugfix
Test: atest TransitionTests#testTransientLaunchWithTranslucentTask
Test: While home is recents. Launch an opaque app.
      Enter recents. Switch to a translucent app.
      Home should be paused.

Change-Id: I0e964088b08f0351cdda354717f78c717cdfe81b