Improve View scrollbar fade scheduling efficiency
When rescheduling a scrollbar fade animation, we only ever postpone the animation.
Take advantage of this by not bothering to remove and re-post scheduled animation.
Rather, let an already-scheduled animation happen at the elapsed time, and re-post itself
to a future time if the scheduled animation start time had since been updated.

This fixes an issue where many View invalidations would cause the same animation to be
cancelled and re-issued.
See the referenced bug for evidence and impact analysis.

Fixes: 419083356
Bug: 415954362
Flag: EXEMPT bugfix
Change-Id: If4bf8664adcc0641209ec5bb849bf7db75ceb347