[Flexiglass] Always include the FooterView in the StackScrollAlgorithm
SSA was skipping to include the FooterView to its visible children while
it was displayed together with the empty view, or if it was the first
child.
This condition was there to skip updates on the footer while it was in a
transitional state, but it caused a more severe bug on Dual Shade, where
it is displayed with the EmptyShadeView by default. By skipping the
FooterView update we left it in a stale position, which was coming from
a previous shade state.

Fixes: 415969173
Test: Check the empty state of the Notifications overlay
Flag: com.android.systemui.scene_container
Change-Id: I5cb5831d8f81c98e27f7f654b813e25dc6d96d47