Gracefully handle failure to connect to configuration service
It's possible that in some cases a process might not have access to the configuration service, for example isolated apps. In these cases we do not want to crash the process instead we processed to set up the ProtoLog client without connecting it to the ProtoLogConfigurationService. The caveat being that we not be able to control the logging to logcat in these processes and we might miss some decoding viewer configs

Bug: 410517697
Test: atest TracingTests
Flag: EXEMPT small change
Change-Id: Idfe22d99cf4d59055a8e5d50fc1d6c4906d2ebcb