Open screen recording dialog using statusbar context
As the screen recording dialog needs to be visible also while the shade is collapsed, it is now using the statusbar context (falling back to the shade context only when the statusbar is not available).

The main issue here was that if the screen recording is started while on the lockscreen, the dialog creation was postponed until after unlocking, and the shade was hidden by then. As the dialog was a child of the shade window, it was hidden too.

Bug: 362719719
Bug: 419577795
Test: ScreenRecordPermissionDialogDelegateTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Iee65f9cb44c9c8267e79a86925eccacfc8331cbb