Chronometer: fix countdown by two seconds
Fixes: 418000654

Flag: EXEMPT Bug Fix

When the chronometer countdown landed on the exact millisecond,
it was rounding to the wrong value. This is because when counting
down, it tended toward the second down rather than rounding, but
if it landed on the exact millisecond, it would round up. To solve
this, the tick is now delayed by 3 milliseconds past the second
rather than 1 millisecond past the second.

Test: local tests, broken test app

Change-Id: I7595f6485515facb08f379eb99ce486e37e8299b