Reset IME requestedVisibility of RICT if imeInputTarget becomes null
When the IME input target changes to null, and the control target is the
RemoteInsetsControlTarget (RICT), we'll reset its requestedVisibility
for the IME to the default visibility (false). If the IME was requested
before, this could have caused inconsistencies when another window, that
has the IME not requested, becomes the input target later.

Fix: 420297992
Test: atest DisplayContentTests#testSetImeInputTargetNullResetsRemoteInsetsControlTargetImeVisibility
Flag: EXEMPT bug fix
Change-Id: Ife1598bb9560da877bafcbff2b66542c68de6d1d