Integrate MinMode with Keyguard
Hide the keyguard UI when the MinMode is in foreground. Also, updated the states in DozeParametert to avoid the keyguard UI appears when the MinMode is enabled.

Flag: com.android.systemui.enable_minmode
Bug: 411746510
Tests: manual testing with a MinMode activity and updated the appropriate unit tests; verified that normal AOD workflow still works.
Change-Id: I05a1d8b056a621926023fbc9c805c406d07bd8e4