[flexiglass] Early return if smartspace isn't enabled
Attempting to bind the views when the plugin for smartspace isn't
present causes a runtime crash. Checking whether it's enabled is the way
to guard against that. The check was forgotten in ag/33448654.

Fix: 421454656
Test: temporarily made the check return false and ran with Flexiglass
on; made sure that there was no crash and that the layout without
smartspace looks ok
Flag: com.android.systemui.scene_container

Change-Id: I7228cbda479f6bfc9866429da754891b5a8ee982