ViewFlipper: add inhibit state to prevent updates, even when started
This internal state avoids a potential race condition in how SystemUI previously paused the ViewFlipper, which could have happened if the `setPaused` call was made before the view received the internal `onAttachedToView` update that would trigger auto-start.  This race was never confirmed, but this refactor should prevent it.

Test: manual with tracing
Bug: 409112153
Flag: EXEMPT refactor
Change-Id: Ie1d76a5eec753dd25b91fb29f7bb5dfed8a7b40c