[AOA FFS Implementation] Enable accessory mode through FunctionFS
Write the USB Descriptors and Strings for a AOA FunctionFS instance.
Accessory ep0 will be used as the control and ep1 and ep2 are added as
bulk endpoints. The writing of descriptors and strings to the AOA
FunctionFS instance is behind a flag.

Bug: 407985367
Bug: 421923887
Flag: android.hardware.usb.flags.enable_aoa_userspace_implementation
Test: Checked that accessory mode is started through userspace
Change-Id: I79e61d77d9c3c22943f4b2fd7bd0f57209ad4c4f