Fix shade header Clock sizing on display move
When used from compose there was no other class propagating the configruation change to the clock. However, the clock receives the configuration itself though the standard View#onConfigurationChanged, so there is no point in propagating it from elsewhere, as it might only result it the wrong signal, and is more error prone.

Also, there was no point for the Clock to implement ConfigurationListener, as it was never registered with any ConfigurationController.

The onLocaleListChanged call was also misleading, as there was no implementation of it in Clock.java.

Bug: 362719719
Bug: 417959771
Test: ShadeHeaderControllerTest + expand/collapse dual shade + move between displays. After the fix clock dimensions are always correct.
Flag: com.android.systemui.shade_window_goes_around
Change-Id: If8ea2a83f4c707a8e8dc17a27862c1301d352197