android.sdk aconfig flags: add android.sdk.device_initial_sdk_int_full
Add a flag to guard the upcoming API Build.DEVICE_INITIAL_SDK_INT_FULL
(to be added in an upcoming CL).

The flag is currently marked as non-exported. This will change when the
flag is used to @FlaggedApi guard the new API, but because the build
doesn't allow exported flags without corresponding API, the flag starts
out as non-exported.

Bug: 419702369
Test: treehugger
Flag: EXEMPT no need to flag the introduction of new flags
Change-Id: If3157d5a59535f9410bc30fd0c2d3fb88cc82cbe