FalsingManager - Re-add existing belief listeners
If there is a change to the FalsingManager, either from a plugin
or a device config change, the prior FalsingManager was getting
cleaned up which included all existing listeners. The result was
all existing listeners were never informed of falsing events,
which could lead to falsing behavior being disabled.

Reattach existing listeners to the new object.

Fixes: 421594970
Test: Ensure falsing continues to operate for long periods
Flag: EXEMPT bugfix
Change-Id: I91e630559971ee898f104269c62c9916961028fc