Autoclick: Center indicator ring around the mouse
This change revamps the ring positioning logic. The previous logic
left the indicator with stale coordinates so when it began drawing
the indicator would not be located on the mouse.

Now the indicator is continosuly updated with mouse coordinates so it
can snapshot its position and use it for drawing.

Demo before: http://b/421980970#comment3
Demo after: http://b/421980970#comment4

Bug: b/421980970
Test: AutoclickControllerTest
Flag: com.android.server.accessibility.enable_autoclick_indicator
Change-Id: Ie699187a0d8ed4865e9f514d37754b25bae54f27