Update PIC Ravenwood dependencies
PIC supports three kinds of nonces: shared memory, sysprop, and local.
Shared memory is used for system nonces when shared memory is
available.  Now that ApplicationSharedMemory is a committed feature of
Android, shared memory is only unavailable when running a Ravenwood
test; this CL makes the selection of nonce type depend explicitly on
whether or not Ravenwood is running.

In particular, if Ravenwood is running then there is a single,
stand-alone process and there is no need for PIC to use either shared
memory or sysprops.  All nonces can use local storage.

Flag: EXEMPT bugfix
Bug: 423886240
Test: atest
 * CtsAppEnumerationTestCases
 * CtsMediaAudioPermissionTestCases
 * CtsMediaAudioTestCases:AudioRecordTest
 * CtsOsTestCases:IpcDataCacheTest
 * FrameworksCoreTests:IpcDataCacheTest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * ServiceBluetoothTests
 * android.app.appops.cts.AppOpsLoggingTest
 * android.content.pm.cts.PackageManagerTest
 * android.media.audio.cts.AudioFocusTest
 * com.android.server.pm.PackageManagerTests
Change-Id: If37c1c1a20d9c1c49aa4dc5fab118601ee716e66