Notify launcher when bubble overflow data has loaded
For the optional overflow, shell needs to notify launcher that
there's stuff in the overflow so it should show. We load the overflow
data from xml and we don't let launcher know when it's loaded.

Instead launcher gets the event whenever it gets the next bubble event
causing the overflow to appear with a delay.

This cl dispatches a change event specifically in the overflow
case, so the overflow shows up when its ready.

Bug: 334175587
Test: atest BubbleDataTest
Test: manual - have stuff in bubble overflow, reboot device, get
               a bubble in bubble bar, observe that overflow is added
               smoothly
Flag: EXEMPT - bugfix / some behind com.android.wm.shell.enable_bubble_bar
Change-Id: I8abc549b99090eee951ff0fc47731a742cb8030d