AudioService: ringMyCar carkit/headphones rings in vibrate
Update the AudioManager.shouldNotificationSoundPlay() behavior
to enforce ringtone playback on "private" devices such as
carkits or BT headphones even when the device is in
vibrate mode.
Availability of such "always ring" devices is evaluated
in AudioDeviceInventory whenever devices are connected/
disconnected, to allow shouldNotificationSoundPlay()
to query that information in a lock-free manner.

Bug: 415311394
Flag: com.android.media.audio.ring_my_car
Test: receive call in vibrate mode with HFP/carkit

Change-Id: Ia72085d496e2785031889c9e844ec363b5c2cc50