Support relative insets in wallpaper service
The wallpaper service didn't provide host bounds when calculating insets
based on an assumption that there should be no caption bar attached to a
wallpaper. However, this is not true on desktop environment. In
settings, when the user opens a wallpaper preview in a freeform window,
there will be a wallpaper service instance running inside a window with
a caption bar.

Provide the window's visible frame to let the insets calculation running
as expected.

Bug: 424504547
Bug: 277292497
Test: Open wallpaper settings - view all - preview and it will not crash
Flag: com.android.window.flags.relative_insets
Change-Id: Ia1a4fd9b1b88a76739876ec9f6169f79ef3a2a10