Clear the drag&drop state on drag end events
In some rare occasions, the drag could end before the drag&drop listeners were notified. This would cause the tile to be wrongfully marked as being dragged until another drag gesture was started.

The fix is to listen for the end event and clear the state is a drag is active at that moment.

Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Fixes: 425555356
Test: manually - Testing really short drag gestures
Change-Id: I31883811e3e18350c91ef63d7071ab1fe60dd766