Stop route scanning after succesfully connecting to route.
Currently, we stop route scan at the time of device discovery and issuing a connection request. This causes the route to disappear during an ongoing connection attempt leading to failure to cast.

The solution is to continue scanning for devices until the connection is established with the suggested device once it is clicked in the UMO.
Additionally, added a check for connection status when reporting connection attempt completing due to connection success vs timeout.

Bug: 422720290
Flag: com.android.media.flags.enable_suggested_device_api
Test: Manual testing and ran InfoMediaManagerTest, LocalMediaManagerTest
Change-Id: I6e7b6933d2aaee5813e050bc596767fe12d1db77