Attempt device entry when WM unilaterally starts the going away animation.
This happens when an activity is launched over insecure keyguard with a flag (or activity option, or command line argument) indicating that the launch should dismiss keyguard if possible.

We were currently ignoring this call in flexi, which meant such activities were either never displayed (if not SHOW_WHEN_LOCKED), or did not dismiss keyguard leading to failures.

Bug: 416718983
Test: atest KeyguardTests
Flag: com.android.systemui.scene_container
Change-Id: Id47b3ce95d96d38fd7a7e73e0ed236e8e8f0b9a7