Resolve compat perf regression via @NoLogging annotation
This annotation allows marking specific compat changes so that they are not logged by the compat framework. This is expected to be used in code that is called in performance critical paths.

Note this change also means that non-loggable changes (including targeting older sdk) are no longer going to stats log.

Change-Id: I3df7d0a3575534fc5be7911624237b57c4deb8f7
Flag: EXEMPT bugfix
Test: PlatformCompatFrameworkTests CtsAppCompatHostTestCases FrameworksServicesTests_android_server_compat
Bug: 424051595