Fixup the seInfo for the restarting services
It's a legacy behavior that the interval between service restarts
might be quite long, while the app could receive updates in between.
If the update is a split APK update, it won't clear the pending
restarts, but update its appInfo. But it'll also clear the seInfo if
that's been set in previous runs (because we'll reuse the same
ServiceRecord instance), this causes incorrect seInfo in the subsequent
runs and it can't be recovered until a force-stop.

For the fix, now we're appending the seInfo during bringing up the
service. We might need to revisit the code to get rid of the service
record reuse behavior for the apk update case, though that's a
relatively larger change and should be done in future releases.

Bug: 424695730
Test: Manual - see b/424695730#comment37
Flag: EXEMPT bugfix
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:be1aa4921c68e109566c9f00cf3c2a4d9154b862)
Merged-In: I25721907e0e10985c1899391cee67f167c04b26b
Change-Id: I25721907e0e10985c1899391cee67f167c04b26b