Consume the modeRejected event based on Event Registration flag value
The modeRejected event sent by Surfaceflinger was consumed by all
the listeners (Display Event Reciever and Chreographer) causing
a crash in Choregrapher for this event since dispatchModeRejected()
in Choreographer is not registered.

This patch adds a Event registration flag corresponding to the
mode rejected event and only the listeners where this flag
is set will recieve this modeRejected event sent by
SF. SO in this case only DER recieves this event.

Test: manual test
Bug: 409640463
Flag: EXEMPT bug fix

Change-Id: I7312baa6c21b2a9ddd160713af5873a5a6b90f7d
Signed-off-by: Manasi Navare <navaremanasi@google.com>