Improve the fling gesture for dividers
Previously, flinging the divider would always send it all the way to the other side of the screen (mFirstSplitTarget and mLastSplitTarget to be precise). With more snap targets on-screen with flexible split, this started to feel janky.

Now, a fling gesture will send it to the closest snap target in the direction of the fling.

Fixes: 428784487
Flag: com.android.wm.shell.enable_flexible_two_app_split
Test: Manual
Change-Id: Ie1b7544c99cdef134023fd11bfae74add2684794