Resolve missing whenToDreamSettings.
This changelist introduces a helper method that can be used
to restrict the whenToDream setting to a set of known values
(usually those presented in settings ui). If the current
setting is not found, dreams are disabled and whenToDream
is set to the default if said value is in the provided list.

Test: atest DreamBackendTest
Bug: 419323985
Flag: com.android.settings.flags.resolve_missing_when_to_dream
Change-Id: Ic234452a5f280ae2cb1ad0e336f343d807523e72