Make most of the windowing modes compatible
When a bubble task, initially in multi-window mode, finishes its
activity and exits its bubble, the task becomes fullscreen
windowingMode (sets as undefined and inherits the fullscreen
windowingMode from its parent). Since the two windowingModes are
not compatible, multiple recent tasks are added if this process
is repeated.

Unless one of them is pinned and the other is not, all windowing
modes are compatible

Bug: 427336494
Test: RecentTasksTest
Flag: com.android.window.flags.fix_task_compatible_modes
Change-Id: Ia8cc8b6ebb73c3dfd4ab2f0d8b1e43bd848f3b2e