Make sure we always call adjustSpacingForImage
The condition for adjusting margins only when not showing the right icon
was actually from an older iteration, and was not correct (leading to
bugs where the spacing looked wrong).

I also made sure we're not fetching resources unnecessarily, since we'll
be calling this method more often.

Fix: 429423146
Flag: android.app.notifications_redesign_templates
Test: manual, screenshot tests should be unchanged
Change-Id: Ife591964a0b83377889d263eed0d23423fa4493b