Fix keyguard notification flicker when transition to dream with shade open
This CL fixes a bug where the notification icons and notifications on
keyguard may show up briefly when transitioning from the lockscreen
with the notification shade open to the dream.

DreamController sends an ACTION_CLOSE_SYSTEM_DIALOGS intent when
starting the dream causing the notification shade to close quickly.
Normally, the KTF transition from LOCKSCREEN to DREAMING explicitly
sets the keyguard alpha to 0 when the transition starts. However the
collapse of the notification shade isn't coordinated with the dream's
launch in any way. This means that the shade collapse might finish
after the KTF transition has already started. This update of the shade
collapse causes the flicker as normally keyguard fades back in when the
shade collapses.

Bug: 409720192
Bug: 430694649
Test: atest SharedNotificationContainerViewModelTest
      also manually verified on device
Flag: com.android.systemui.lockscreen_shade_to_dream_transition_fix
Change-Id: Idf4a1911acf5571ffda4cd6b133958e478f7f71d