[Dual Shade] Gracefully switch from Single shade to Dual shade.
It possible that while Single shade is shown, shadeMode abruptly changes
to Dual. For example, when rotating a phone from portrait to landscape,
or when unfolding a folded device.

The reverse transition from Dual to Single shade is equally possible as
well, of course.

We therefore add the ability to recognize and gracefully switch scenes
in such a scenario.

Fix: 430084116
Bug: 411365508
Test: Tested manually by rotating a phone with single shade enabled, and
 by unfolding a folded device.
Test: Added unit tests.
Flag: com.android.systemui.scene_container
Change-Id: I4e50012173e8d4bd17f4108f1a343d55d471aa5a