[85/n] Implement TaskIdResolver
When dealing with MultiWindow (e.g. split screen) the id of
the task received in a Change might not be the id of the task
associated to the related letterbox surfaces.

To handle these cases it's useful to store the parentId which
will be used to detect the right task id.

Flag: com.android.window.flags.app_compat_refactoring_fix_multiwindow_task_hierarchy
Bug: 430486865
Test: atest WMShellUnitTests:TaskIdResolverTest

Change-Id: I29e1f3a49c6352eba93143cedbe002b06915f8ff