Fix transition when pressing power button from hub gesture
This CL makes the hub snap immediately to blank when the devices goes
to sleep. For a single power press, this happens after the screen is
off so there's no visual issue. With a double power press that's slow
enough to not trigger camera, the device ends up on the lock screen as
expected.

Prior to this change, the KTF transition wouldn't finish until the hub
STL became idle, leaving the device in DOZING but with messed up UI.

Bug: 431931326
Fixed: 431931326
Test: atest KeyguardTransitionScenariosTest
      manually verified on device, see bug for video
Flag: com.android.systemui.communal_power_transition_fix
Change-Id: I3c7e116bf1a9efc8309893492cf162fc3e4a6e1f