Replace long click handler with long click semantic indicator.
The long click handler was not doing anything because long clicks on ExpandableNotificationRow are not handled by simply calling view.performLongClick. Adding actual long click handling to BundleHeader was therefore sometimes consuming the long click that would have triggered opening the guts, without actually doing anything. It was happening "sometimes" because the long-click timeout was probably close enough between the two handlers that it was a toss-up on which handler activated first.

This puts us back in the state where BundleHeader is handling clicks and ExpandableNotificationRow handling long clicks, but with an extra label. It is still the case that both elements are selectable accessibility nodes for things like Voice Access, which can be ambiguous.

Fixes: 430881139
Test: manual
  - repeated long press on the bundle header always opens guts
  - confirmed header is listed as long-clickable in TalkBack
  - performed long click in TalkBack, Voice Access, Switch Access
Flag: com.android.systemui.notification_bundle_ui
Change-Id: Ia2f4d43745f096d723c522f448bcc4d64ab68d14