[flexiglass] Always compose QS/shade scenes and overlays.
This is a performance optimization meant to reduce overall jank when
transitioning between scenes and overlays. The content for the QS and notification scenes and ovelays gets drawn all the time but only
the current scene and overlays become visible.

Bug: 430488243
Flag: com.android.systemui.scene_container
Test: manually verified I can transition between all scenes normally
with single and with dual shade

Change-Id: Ie6f4c6a9be8868eb2153719f5ed3332d45c88469