Properly propagate rotation onDisplayChange
This CL does a couple of things:
  - when we receive IDisplayChangeWindowController remote
display change from Core, we should allow rotation changes
to be propagated into updating display layouts too.
  - when updateDisplayLayout() is run, if rotation has happened
along with display size change, the dimensions of endBounds are
already in the new orientation. So we should rotate the local
displayLayout before resizing it.

Bug: 432725672
Flag: EXEMPT bugfix
Test: manually repro the steps in the bug
Change-Id: I9c361609a9df896029026a155340e3bda440265f