Fix bug where dim layers would disappear on taskbar tap
When the user touches the Taskbar, a transition is started in preparation of a potential swipe up to recents. In splitscreen, this reparents all the app surfaces, the divider, and the dim layers onto a transition/animation root.

Previously, dim layers were not properly ordered above the app layers on this new root, causing them to disappear as soon as they were reparented onto the transition root.

With this change, we use a new approach where the dim layers are reparented onto the app layers instead, and have their layer set to maximum there. This solves the layering problem, and additionally, going forward, the dim layers will always inherit the app layer's crop and position, instead of having to be set separately.

Fixes: 431036371
Flag: com.android.wm.shell.enable_flexible_two_app_split
Test: Visually correct when a 90:10 pair goes to recents and back
Change-Id: I9c50b9516fb55adacad81055d86c0ec4135cde18