[1/n] Round to the nearest integer after all division operations in the bounds calculation
Currently, we are adding +0.5f to round to the nearest integer in bounds calculation. However, there are 2 division operations that are missing the addition of +0.5f. This causes test failures, especially when we update the default initial bounds scale from 75% to 72%

This CL fixes the issue by adding +0.5f margin to these left cases, which prepares the ground for updating to 72% scale. We also add the same +0.5f margin to 2 test cases to prevent test failures by only one unit difference, which occurs due to `float` to `int` casts after the division operation in the tests.

Flag: EXEMPT calculation fix
Test: atest WmTests:DesktopModeLaunchParamsModifierTests
Bug: 433505659
Change-Id: I8bcb65efe628b439ccc4cdebae8d9d9a6368a18f