Split root task should not allow bounds for child tasks
Call the new WindowContainerTransaction API
setDisableChildTaskBoundsOverride when split root task is ready.
Calling this on the root task ensures that when a task is reparented
under the split root task, the child task bounds are cleared.
Split screen requires that the tasks inherit the bounds of the root
tasks, if task has bounds set on it, it won't pick up the split bounds,
resulting in incorrect placement on screen.

Bug: 427245857
Test: atest WMShellUnitTests:StageCoordinatorTests
Flag: com.android.wm.shell.split_disable_child_task_bounds

Change-Id: Icd2bbc10d7b2c9398c95040af74e891ad6393cee