[Flexiglass] Remove the extraneous HUN placeholder from the LS
HeadsUpNotificationPlaceholder was included twice on the Lockscreen when
bypass was unabled. This lead to a crash, because the LS scene was
trying to compose the same shared element twice.

This CL removes the extra definition for HeadsUpNotificationPlaceholder,
because the Notification section already contains one, which we can use
on the bypass LS.

Fixes: 432246432
Test: Unlock the device with face auth bypass
Test: Receive a HUN on the bypass LS
Flag: com.android.systemui.scene_container
Change-Id: I7bafd4f81df4d7f786d87cd5a8807ba2a93cf55a