[Flexiglass Dual Shade] Handle taps on status bar
Only mouse click is handled before. This change addtionally
handles single tap gestures.

A GestureDetector is used to reliably distinguish a tap from a
swipe. This prevents the view from consuming swipe-down gestures,
ensuring that the main gesture to open the notification shade continues
to work correctly.

Also added a config to guard this feature for desktop only.
This config will be overriden to true for desktop in the next cl.

Bug: 429237899
Flag: com.android.systemui.scene_container
Test: PhoneStatusBarViewControllerTest
Change-Id: Id530bc08688181cc69b04d1d9d116ba5eef9b5c5