Fix correct scrim visibility not dispatched when closing bouncer on dream
The bouncer to dream keyguard transition manually fades the behind
scrim for a smooth transition. This can cause issues as the shade
window stays visible as long as any scrim is visible, but
ScrimController itself stops dispatching the scrim visibility state
when the scrim transition ends. This is not connected to the KTF
transition, meaning if the keyguard transition lasts longer, the scrims
are still visible by the end of the scrim transition.

This CL manually dispatches the scrim visibility state at the end of
the KTF transition.

Note that the scrim_fix flag prevents the issue of the shade window
staying visible in general. Still proposing this fix so that the scrim
visibility state (as seen by the shade) remains accurate.

This change introduces a new flag to control the bouncer transition fix
as it should not be rolled out without this fix as well.

Bug: 409720192
Bug: 434038809
Fixes: 434038809
Test: atest ScrimControllerTest
Flag: com.android.systemui.dream_bouncer_transition_fix
Change-Id: I8006c9f8aef139e6386f0cbadcf7eb6a78e0c584