Set protectionLevel to module for Bluetooth permission
Bluetooth used to get some privileged permission thanks to its role. But
there are two problem with role given permission:

1/ During a normal boot, there is a race between Bluetooth starting it's
app process, and the permission module, trying to give the role to
Bluetooth. If Bluetooth start before, the role will kill the app and
Bluetooth will not be notified of the issue because it may not be bind
yet.

2/ On any devices, the role is given thanks to a package that is
determined by the config. The package name can change if this is a
mainline module or not. By removing the role we reduce the number of
things that depend on it

Bug: 395709071
Test: flash & boot
Flag: android.permission.flags.bluetooth_permission_use_module_protection_level
Change-Id: I43185ab7d7e2122f1f7c72cf8969402d75abb4c0