[flexiglass] Simplify ShadeMode + remove ShadeRepository.isLargeScreen.
Now that isShadeLayoutWide was split into "isWideScreen" and
"legacyUseSplitShade", which differ by the fact that the latter is
"false" on large screen devices in portrait orientation, we no longer
need the "isLargeScreen" flow in ShadeRepository.

This is part of a multi-CL refactor that is meant to disambiguate the
meaning of "shade layout wide" and simplify related decision-making
logic.

Bug: 354926927
Bug: 411365508
Test: All unit tests pass.
Flag: com.android.systemui.scene_container
Change-Id: I9b356ff7e31b8cf461fa5406cd422b9a06ce9fbe