Fix failing wallpaper B&R GTS tests
Make sure the device info file is always backed up when it doesn't
exist, even if the shared preferences are up to date. Otherwise, the
device info file is not backed up in GTS tests, leading to some errors.

Also make sure to properly back up when WallpaperManager APIs taking a
single "Rect visibleCropHint" are used. In this case, do similarly as
before: adjust that "visibleCropHint" for the new device based on the
largest non-rotated screen on both devices, so that what's on the new
largest screen keeps the same center as before.

Flag: android.app.fix_wallpaper_crops_on_restore
Bug: 332937943
Bug: 435027053
Bug: 435027428
Test: ./vendor/xts/tools/gts-tradefed/etc/gts-tradefed run  gts -m GtsBackupTestCases -t com.google.android.gts.backup.WallpaperBackupTest

Change-Id: I448dae0e60922b9c749dca285962bd17d1588043