Fix failing UserManagerServiceMockedTest related to admin flag
In this test, whether the system user has FLAG_ADMIN is determined before setSystemUserHeadless() is called.
If run on HSUM device, the system user does not have FLAG_ADMIN, which caused a few tests to fail.
So setSystemUserHeadless() needs to set the flag correctly, depending on the feature flag hsu_not_admin.

Bug: 436931087
Flag: EXEMPT test only
Test: atest FrameworksMockingServicesTests:com.android.server.pm.UserManagerServiceMockedTest

Change-Id: I658c8a9bfa2f51867a2b41ea8d8241e647f6a34d