BouncyBall: Specify "relevant frames" for automation
Our app can run "forever", but for the purposes of automated testing,
we obviously need to keep this finite.

We already had a concept of things being "too early" in the app.
We now add the idea of our test time (2 minutes).

We use a trace counter to note which state we're in, so automation
can use that information in a trace to only analyze the relevant
frames.

Bug: 408044970
Test: We're able to see and use this state appropriately in automated trace analysis
Flag: EXEMPT for test app
Change-Id: Id9c0738094a4d984d26c56bcd6abba37bbaa3dc6