Add test for move-to-next-display in projected mode
This CL adds an end-to-end test for the move-to-next-display
keyboard shortcut when the device is in projected mode.

The test verifies that a task can be moved from the internal display
to an external display, and back again.

Bug: 435069150
Test: MoveToNextDisplayProjectedTest
Flag: com.android.window.flags.move_to_next_display_shortcut_with_projected_mode
Change-Id: I463cb4619bb5933b83990928f5e66fdec5b9e370