Ensure that device entry does not always dismiss the current scene.
If a back stack exists, device entry might mean "remove the
lockscreen but keep the shade open". By updating the back stack
instead of simply moving to `Scenes.Gone`, we enable this behavior.
Any `OnDismissAction` triggered after device entry can still decide
whether the shade should be dismissed or not in the specific
instance.

Note that this does not perfectly match the pre-scene container
behavior yet, but that's because of a different issue with the scene
background. What's happening underneath is correct.

Fix: 436602773
Flag: com.android.systemui.scene_container
Test: atest DeviceEntryInteractorTest SceneBackInteractorTest SceneContainerStartableTest + manual (see video in the bug)
Change-Id: I27f0baa80ec434ff040cd64099c38792c48a6e38