PackageManagerNative: Include app's sourceDir
Include the sourceDir of the application (if present) in the result of
getPackageInfoWithSigningInfo.

Additionally, add jajar rule to rename the exported class by the jar to
reduce possibility of multiple jars exporting this same symbol.

Requirement: virtualizationmachine service which uses native pm consumes
VM configuration supposed to contain the packagename of the apk. It
requires sourceDir of the apk to mount the package into Microdroid VM.

Test: adb shell service call package_native 2 s16 "com.android.microdroid.test" i32 0
Test: Compare above with `adb shell pm path com.android.microdroid.test`
Bug: 438179019
Flag: build.RELEASE_AVF_ENABLE_ADVANCE_MULTITENANCY
Change-Id: I91b31b9bf59bc970b6e262d7386ef8954efb2c84