Remove uptime check for crash rate limiting
When the network stack process is lost and ConnectivityModuleConnector
skips crashing, other system components generally crash shortly after
(for example NsdService trying to use tethering callbacks, or wifi
trying to communicate with IpClient). The only way to recover from the
bad state is to crash, so make sure it happens even in the first 30
minutes after boot.

This makes sure the system recovers faster, and should make it easier
to investigate and fix issues causing the network stack process crash,
as it avoids reports coming from other components.

The "at most once every 6 hours" rate limit is still present (for now),
so this should still avoid bootloops. Reducing or removing that limit
may be considered if it turns out the system is still left in a bad
state too often (causing reports of crashes in other components).

Bug: 434820757
Test: m
Flag: EXEMPT bugfix
Change-Id: I344052e1614f39c295b8c0701f46871e8c6491a2