Reapply "Fix flicker when swiping to shade or QS"
This reverts commit cd918d2b9da9a229d060d32a3563aefbda2c8b14.

This CL reapplies ag/35037546, which was incorrectly using the nested
STL scope in the QuickSettings scene instead of using the outer one
(see b/437888180#comment12).

To fix this, we are now exposing a
ContentScope.isAlwaysComposedContentVisible() boolean that also takes
ancestors into account.

I made sure that the tests that failed in b/438423768 now pass with this
CL.

Bug: 437888180
Bug: 438423768
Test: android.platform.test.scenario.sysui.notification.NotificationShadeChrome#testScrollingToBottomRevealsFooter
Flag: com.android.systemui.scene_container
Change-Id: Icde53b7479bd559100fca29038592fac286b7ff9