[flexiglass] Fixes DeviceEntryHapticsInteractorTest
A race condition between when the class under test was instantiated
(pulling in all of its dependencies) and when, later, the biometric face
auth bypass was set as a config resource override caused the bug.

The fix is to delay instantiation of the class under test and to
modernize the test to use Kosmos.runTest.

Fix: 437436426
Test: test passes when run both instrumented and deviceless
Flag: EXEMPT test fix only
Change-Id: I75b0c41dc6cd4cbf464855c45a10fe5b5bd1acec