Enter desktop before setting up external display in DragMoveWindowToNextDisplayTest
Currently we setup the external display then launch the test app in
desktop. But it seems the app launch happens too quickly after the
display setup is triggered which does not given enough time for the
display window mode to take affect. This results in the app lauching
will fullscreen bounds in freeform.

Instead launch the test app first and enter desktop before triggering
the external display setup. This allows the test app to enter desktop
succefully with the correct bounds.

Flag: EXEMPT test fix
Test: atest DragMoveWindowToNextDisplayTest
Bug: 432227275
Change-Id: I3c09c9e1670f4ca1f1dded07cf0f8e8a0763abf7