Add a read only flag library for resource flags
This is because I am working on a change ag/34724936 that should be
flagged in native code. The change is in libandroidfw which aapt2
depends on. Aapt2 has a windows variant while the flag libraries do not
so we need a read only version of the library. The flag itself (and
almost all resource flags) is read only and aapt2 itself will never
depend on the flag being enabled so this is fine.

Test: N/A
Bug: 420720496
Flag: EXEMPT can't flag bp changes
Change-Id: Ia1e0f8e9eb01a4945ad246321a75135e4bc202a6