Initial drop of the memory test app
This is a bare-bones application that is used to measure the baseline
memory used by an application when it starts.  The baseline memory is
memory that is not under control of the application.

The test is run via atest and one of the test artefacts is a perfetto
trace that contains a heap profile.

Flag: EXEMPT test-only
Bug: 429215212
Test: atest
 * AppMemoryTestCases

Change-Id: I8ef0435e818ced380b5bb763285520a587a12056