Do not crash when root appears without matching request
It's possible that a display becomes inelegible for desktop while a desk
root task (or minimization root) was being requested, and that the root
created is reparented to another display when this happens. In such
cases, the task may appear (in ShellTaskOrg) for the first time under
the new display instead of the one that originally requested.

With this change we no longer crash SystemUI and log an error instead.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Fix: 439173763
Test: atest WMShellUnitTests
Change-Id: I7846a447cfa78e1f653ba49a3cd24a51bfa3c617