[Flexiglass] Fix FooterView flicker when expanding/collapsing DualShade
During the expansion of the DualShade panel, the height of the stack and
the draw bounds are animated synchronously, but when we reach the
overshoot effect, there is a point where the StackScrollAlgorithm thinks
that there is no space for the FooterView. This is causing a visual
flicker.

Fix this by ignoring _SSA noFooterView_ updates until an expansion is in
progress.

Fixes: 436760595
Test: open DualShade by scrolling, with a fling, fastly, slowly, etc.
Flag: com.android.systemui.scene_container
Change-Id: I869171684bc5c8f0e35c00794048a8c6255d694d