Allow only StatusBarTouchShadeDisplayPolicy for shade display
This removes the option to use any other ShadeDisplayPolicy: the shade from now on will only be able to move based on the latest statusbar interaction.

Still not deleting the other policies as they might be useful for testing and experimenting with different behaviours.

Bug: 439820939
Test: MoveShadeOnStatusBarTouch
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I126c4a165da66003047d1e3034957f1872b25cb6