Count root tasks rather than windows
This CL changes cuj5e/p to count root tasks instead of windows.
The test cases use Chrome browser as a test app but Chrome may open a
popup upon its launch (e.g., Ad privacy check), which unexpectedly
increases the number of windows for the app.
What the test really wants to verify is that there is only one app
instance for the app. So counting a number of tasks should be more
appropriate.

Flag: EXEMPT - test fix
Bug: 440202064
Test: ConnectedDisplayCujSmokeTests
Change-Id: I6210f48af4f58f3866016d2acd49b799042ce4f3