Fix BackupManagerConstants observation for non-system users
On devices with Headless System User Mode (HSUM) enabled, BackupManagerConstants was not being notified of settings changes for non-system users.
This was because it was initialized with the system user's ContentResolver, causing it to observe settings changes only for the system user.

This change addresses the issue by creating a user-specific Context within UserBackupManagerService for the relevant user ID.
The ContentResolver from this user-specific context is then used to instantiate BackupManagerConstants, ensuring that it correctly observes settings changes for the intended user.

Bug: 413811693
Bug: 415890568
Test: atest android.cts.backup.SuccessNotificationHostSideTest
Flag: EXEMPT bug fix
Change-Id: Ic5c5427e1746032446aaab27d41420f21458e3ca