Fail duplicated attachment to dream service
Previously if the same dream service is attached a second time, it fails
silently, leading the DreamController to lose track of the previous
dream, and not stopping it properly. This change makes the service
communicate attachment failure back to the DreamController so that it
can manage its state properly.

Test: atest DreamControllerTest
Test: verified the failing CUJ listed in the bug, and dream stops when
      attached a second time, intead of getting stuck
Bug: 434674824
Flag: android.service.dreams.allow_dream_attach_failure
Change-Id: I47789eb618cffc7bb229a47974e450dfc3035b1a