Mixed transitions can enter PiP if scheduled only
Having a change with a windowing mode = pinned isn't
enough in a mixed transition to deem its part to be
animating enter-pip.

More specifically, if a mixed transition wants to check
whether to animate enter-pip, it needs to check whether
PipTransition sent a hierarchy-op to Core to actually move
a PiP activity into a pinned task. This happens if and only if
PipTransitionState#getState() has entered SCHEDULED_ENTER_PIP.

Bug: 432014880
Flag: EXEMPT bugfix
Test: manually repro the steps in the BR
Change-Id: I7268aa76880741db0d8d930bfacf5e20204850bb