Moved broadcast receiver blocking call to the background thread.
According to documentation :

Executes the given runnable. If the caller is running on the same looper
as this executor, the runnable must be executed immediately.

The caller was a UI thread, so the method was executed blocking the main
thread.

Introduced change that force-post the runnable to be executed later.

Fixes: 441416105
Flag: EXEMPT bugfix
Test: TreeHugger
Add log after changed line to check that broadcast execution completes
before collapseStack() is invoked in the main thread.

Change-Id: I6f4f292cd8e6a880721b31fcf8e11fb8d3546765