Safety check for OCCLUDED on wakeup
There is a rare case where on device wakeup, occlusion also starts
as during a phone call. If these signals come in at just the wrong
time, the existing FromAodTransitionInteractor will decide to go
to LOCKSCREEN. This adds a failsafe check to ensure any transition
to LOCKSCREEN gets redirected to OCCLUDED if need be.

Fixes: 441275202
Test: atest FromLockscreenTransitionInteractor
Flag: EXEMPT bugfix
Change-Id: Ie88f7c6399911b5859e1c70c601deb884499f7f6