Add isParentGroupExpanded method to ExpandableNotificationRow
There are parts of NotificationContentView that used isGroupExpanded() to determine if a notification is a child in an unexpanded group, in which case it should be displayed as a single line and its (min) height is set accordingly. This is incorrect in the case of nested groups (for instance, a group inside a bundle), where isGroupExpanded() checks the expansion state of the group itself and not its parent.

Bug: 440448363
Test: manual, by turning off "show sensitive content" for notifications on the lock screen and then verifying that the notifications show at the expected height
Flag: EXEMPT bugfix
Change-Id: Id5e7e99b4c9d69ea6a4e4194d415a51ff9f9d72d