Ignore wallpaper task when moving task to next display
When moving a task to the next display, the move is skipped if the
target display has a focused non-desktop task. The desktop wallpaper can
be considered a focused task, which prevents moving a task to an empty
desktop.

This change filters out the wallpaper task from this check, allowing
tasks to be moved to a display that is only showing the wallpaper.

Test: DesktopTasksControllerTest
Flag: com.android.window.flags.enable_move_to_next_display_shortcut
Bug: 441791586
Change-Id: I44579b207d6d6cbf13baf28a1f77bf14c749c0f1