Workaround missing display leashes
There is some sequencing problem where display leashes aren't
present in time to use them for transitions on that display.

When TransitionHandlers were reparenting to the null display leashes,
that would be treated as just moving them offscreen. However,
relZ-reparenting to null is not allowed and so this recently became
the source of crashes during CTS tests.

Raise a Log.wtf for investigation, but don't crash the system ui process
any more.

Test: None
Bug: 420068074
Flag: EXEMPT BUGFIX
Change-Id: I8bb6307590dcc73f08f48a2d7799fa09a45c9b49