Deprecate canAddPrivateProfile
Marking this method as deprecated in preparation for its removal. We
cannot remove it yet because Tradefed uses it, and we cannot update
Tradefed yet due to trunk stable flag issues. But in the meantime, we
want people to use know that they should use the standardized
canAddMoreProfilesToUser instead, and not copy this overly specialized method.

Bug: 413464199
Test: N/A -  just documentation
Flag: android.multiuser.consistent_max_users
Change-Id: I7998630b0bef44efedf10dcddcb19d0ad2c67bc7