[Flexiglass] Add padding between Media and Notifications on Lockscreen
The vertical spacing between these elements used to be correct, because
NSSL was always adding a padding before its first element due to a bug.
With this bug being fixed and NSSL positioning its children to the top
again, the spacing disappeared.

This CL fixes it by adding a spacing between the Media and Notification
elements in Compose.

Fixes: 443676503
Test: check the Narrow LS with Media and Notifications
Test: check DualShade LS on large screens with Media and Notifications
Flag: com.android.systemui.scene_container
Change-Id: I37c572d7e9baf1eb5a6fb1519171d677f47a7800