Refine USB data protection notification silencing and logging.
This change fixes when USB data protection notifications are suppressed due to user-set silence flags, a silent interaction is logged. The logic for setting silence flags in the broadcast receiver is also simplified. Tests are updated to verify the content of different notification types and confirm that silent interactions are logged when notifications are suppressed.

Bug: 439009345
Test: atest, manual testing
Flag: android.security.aapm_feature_usb_data_protection
Change-Id: I1919cdd96e1cdf3a3fbe80a13fd7d48fe86c8df8