Fix calculation of the caption's customizableRegion
Prior to this change, the customizable region was calculated by
WindowDecoration and CaptionController by taking the difference of the
caption rect (in display coords) minus the bounding rects (in tasks
coords), which resulted in the bounding rects rarely actually being
excluded from the customizable region because the task is never near
display's (0,0).

Instead, the bounding rects need to be offset to align with the task's
position in the display before taking the difference.

This change also moves all of these region calculations into a helper
class to reuse with wd1 and wd2 and to make it testable.

Flag: EXEMPT bugfix
Bug: 414521306
Test: atest WMShellUnitTests
Change-Id: I45c7cd98e8589c5fd528437d0f1337a730c35a48