Use a separate ScreenSize class in VirtualInputDeviceController
Before this CL, width and height of the screen were being passed
directly to the low-level function.

For some input device types, it is required to provide valid values for
these, while for others (like upcoming gamepad), it's not needed.

These values are now being put into a separate struct, so that the
calling code can be more explicit that it's providing it (or not).

Bug: 439992405
Test: atest InputTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ided8b0789e9a400beb5362443618e727e7329bdc