Fix NPE from ClipboardService using VDM
Apps that run on the default device and request the clipboard from a
virtual device will fallback to the default device clipboard (and not
crash the ClipboardService).
Considered returning DEVICE_ID_INVALID if the requested virtual device
doesn't share the clipboard with the default context, but preferred to align with the next case where the fallback deviceId is the deviceId of the app that requested the clipboard (which for the case is the default).

Bug: 443849134
Test: atest ClipboardManagerTest ClipboardManagerListenerTest StreamedAppClipboardTest
Flag: EXEMPT BUGFIX
Change-Id: I862c7c5a020edcf8b98b6e3966ca549151bd29d3