Close and remove HubEndpointSession before invoking onSessionClosed callback.
This change ensures that the `HubEndpointSession` is marked as closed and removed from the active sessions map before the `mLifecycleCallback.onSessionClosed` is executed. This prevents a race condition where a client potentially re-opens a session with the same ID within the onSessionClosed callback, which can be removed from `mActiveSessions` in the existing code.

Bug: 443346395
Flag: EXEMPT bug fix
Test: Presubmits
Change-Id: I7917fdd4fe6718ec0c142c05b4f2e82b6dabc394