Fix crash when reading isDreamInPreviewMode
Prior to this fix, callbacks could still trigger between onEndDream,
which is when isDreamInPreviewMode is no longer valid to call, and the
overlay exit animations finishing. This CL adds a new ended state that
is checked before executing code that relies on a connected dream See
bug for a more detailed description of the issue and fix.

Bug: 432039567
Fixes: 432039567
Test: atest DreamOverlayServiceTest
Flag: android.service.dreams.dream_overlay_started_fix
Change-Id: Ib04af1508df08507737a490af1d0f6ee33d0473b