Set windowing mode in options when launching in fullscreen
When we are launchign a task in fullscreen is stage coordinator, also
set the launch windowing mode.
Launch may bring an existing task to front, which may not currently be
in fullscreen windowing mode (for example it is in a bubble).

Bug: 427103183
Test: atest WMShellUnitTests:StageCoordinatorTests
Flag: com.android.wm.shell.split_to_full_set_window_mode
Change-Id: I4a3f8585d6037043e84bfc29db3d8032cd831090